requirements.txt
google-cloud-translate>=3.15.0
click>=8.0.0
python-dotenv>=1.0.0
